% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo2topo.R
\name{geo2topo_str}
\alias{geo2topo_str}
\title{Convert spatial data to TopoJSON}
\usage{
geo2topo_str(geodata, layer, ..., verbose = getOption("verbose"))
}
\description{
Convert spatial data to TopoJSON
}
\details{
Relies on Node: `geo2topo`
}
\examples{
library(sf)
library(leaflet)
library(TIGER2015)
CA_counties_sf <- st_as_sf(TIGER2015::TIGER2015_CA_counties)
topojson_str <- geo2topo_str(CA_counties_sf)
leaflet() \%>\% addProviderTiles("Stamen.TonerLite") \%>\% addTopoJSON(topojson_str)

}
